refactor: MeetingMember 패키지 이동
